Initial attempt
- made 4 dockerfiles (1 standard Dockerfile that doesn’t have any
environment variables defined at all and require env variables to be
passed to the docker run command upon running the container. Then 3
other dockerfiles were made (1 for production, 1 for local, and 1 for
test container).

- Updated the config.yml to use the new Dockerfiles
- updated the docker:run script in package.json to use correct
environment variable.

Do we want production version to run the environment to have the
environment variables in the header too???

Do we want the docker version of the local frontend to run on the same
port as the local developer version of the app? (3000? OR 8080?)